Modified the warmup subcircuit
To initialize the circuit, a subcircuit was used that on powerup would
temporarily add voltage to the + leg of the op amp. Now this circuit adds
this voltage directly to the base of Q1. With this I could also remove
the resistor conecting both cricuits (Rs1).

I also decreased the fill zones tolerances and most of the board is now
filled.