Set new values for components
One of the things I noticed with the circuit is an asymmetry when increasing wind speed vs decreasing. When increasing, the response is fast. But when decreasing it is very low.

This is to be expected: the electronics respond much faster than the cooling: the maximum colling rate happens when there is no current and therefore heat transfer within the thermistor is the slow part.

To make response more symmetric, I increased the the resistors at the op amp input from 1k to 47k. This introduced a new problem: the capacitor would not charge fast enough for the initial charge subcircuit. I could increase the capacitoor in this subcircuit and increase the resistance as well. I haven't tested these solutions. What I did was increase the feedback capacitance from 100nF to 1uF. The circuit is working and the response is much more symmetric.

Another change was decreasing the base resistors. They are not necessary. I still left a low resistor just to keep the layout. But they could be replaced by wires (that's what I tested). This improves response time (which doesn't really matter).