Replaced transistor KSP2222A for BC517
The BC517 is a darlington transistor and the response of the anemometer is much faster (when increasing wind speed). Why would I care about that if I went to all the trouble to make the response more symmetric?

It really depends on the objective. If I want to measure the mean speed in a turbulent flow, using the more symmetric response is better. But If I want to estimate wind gusts, the faster response time could do that! This change inm transistor really has no downsides.