I found out the trouble: too much power dissipated on the tranistor
During normal operation, the current is low and therefore the voltage at the top of the bridge is low and the transistor needs to reduce a lot of voltage and dissipate a lot of heat. After normal operation for a little time the transistor would burn up.

I see 2 solutions:
 1. Replace the transistor
 2. Modify the resistors

I think I will do #2 anyway and do #2.