Replaced the transistor from BC517 to TIP120 and resistors on the bridge
The resistors on the top of the bridge now have the same value as on the bottom of the bridge. This decreased the power dissipated in the transistor.

The new transistor can also include a heat sink dependeing on application. The setup is working on the protoboard. I has been working for the last hour, voltage input can be decreased to 15V and the system will still start. At 13V it takes a while, perhaps a larger capacitor for startup procedure will help.